requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.0
cryptography==36.0.2
defusedxml==0.7.1
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==2.2.5
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.0
fpdf2==2.8.3
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1

